Move matchmaking into businesslogic
