+++ aesops/routes.py
@@ -17,11 +17,11 @@ from data_models.tournaments import Tournament
 from data_models.users import User
 import aesops.business_logic.elim_match as e_logic
 import aesops.business_logic.match as m_logic
+import aesops.business_logic.matchmaking as mm
 import aesops.business_logic.players as p_logic
 import aesops.business_logic.tournament as t_logic
 import aesops.business_logic.users as u_logic
-import pairing.matchmaking as mm
-from pairing.matchmaking import PairingException
+from aesops.business_logic.matchmaking import PairingException
 from aesops.utility import (
     rank_tables,
     get_faction,
