+++ tests/simulation_utils.py
@@ -3,7 +3,7 @@ from string import ascii_uppercase
 from data_models.tournaments import Tournament
 import aesops.business_logic.match as m_logic
 import aesops.business_logic.tournament as t_logic
-from pairing.matchmaking import pair_round
+from aesops.business_logic.matchmaking import pair_round
 from data_models.model_store import db
 import tqdm
